feat:exception handling done

+++ Backend/LostAndFound/src/main/java/com/lostandfound/LostAndFound/Item/rest/ItemController.java
@@ -4,14 +4,12 @@ import com.lostandfound.LostAndFound.Item.entities.Item;
 import com.lostandfound.LostAndFound.Item.service.IItemService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-
 import java.util.List;
 import java.util.Optional;
 
 @RestController
-@RequestMapping(path = "item")
+@RequestMapping(path = "/item")
 public class ItemController {
-
 
     //<editor-fold desc="Private members">
     @Autowired
@@ -29,9 +27,20 @@ public class ItemController {
      * @param id of the item to be retrieved
      * @return an Optional containing the item, or an empty Optional if the item does not exist
      */
-    @GetMapping(path = "{id}")
-    public Optional<Item> get(@PathVariable("id") String id) {
-        return this.iItemService.get(id);
+    @GetMapping(path = "/{id}")
+    public Object get(@PathVariable("id") String id) {
+        try {
+            if(this.iItemService.isItemExist(id)) {
+                return this.iItemService.get(id);
+            }
+            else {
+                return "item could not be found";
+            }
+        }
+        catch (Exception e)
+        {
+            return e.getMessage();
+        }
     }
 
     /**
@@ -39,9 +48,14 @@ public class ItemController {
      * @param isFoundItem true if looking for found items, false if looking for lost items
      * @return a list of items based on the specified condition
      */
-    @GetMapping("get-list")
-    public List<Item> getList(@RequestParam(required = true) boolean isFoundItem, @RequestParam(required = true) double longitude, @RequestParam(required = true) double latitude, @RequestParam(required = true) double distance) {
-        return this.iItemService.getList(isFoundItem, longitude, latitude, distance);
+    @GetMapping("/get-list")
+    public Object getList(@RequestParam(required = true) boolean isFoundItem, @RequestParam(required = true) double longitude, @RequestParam(required = true) double latitude, @RequestParam(required = true) double distance) {
+        try {
+            return this.iItemService.getList(isFoundItem, longitude, latitude, distance);
+        }
+        catch (Exception e) {
+            return e.getMessage();
+        }
     }
     //</editor-fold>
 
@@ -52,8 +66,19 @@ public class ItemController {
      * @param item that should be updated
      */
     @PutMapping
-    public void update(@RequestBody Item item) {
-        this.iItemService.update(item);
+    public String update(@RequestBody Item item) {
+        try {
+            if(this.iItemService.isItemExist(item.getId())) {
+                this.iItemService.update(item);
+                return "Item updated successfully";
+            }
+            else {
+                return "item could not be found";
+            }
+        }
+        catch(Exception e){
+            return e.getMessage();
+        }
     }
     //</editor-fold>
 
@@ -64,21 +89,37 @@ public class ItemController {
      * @param item that should be inserted
      */
     @PostMapping
-    public void insertNewItem(@RequestBody Item item) {
+    public String insertNewItem(@RequestBody Item item) {
         // Create item
-        this.iItemService.create(item);
+        try {
+            this.iItemService.create(item);
+            return "Item Inserted Successfully";
+        }
+        catch(Exception e) {
+            return e.getMessage();
+        }
     }
     //</editor-fold>
 
-    //<editor-fold desc="Get requests">
+    //<editor-fold desc="Delete requests">
 
     /**
      * Deletes an item by its ID.
      * @param id of the item to be deleted
      */
-    @DeleteMapping(path = "{id}")
-    public void deleteById(@PathVariable("id") String id) {
-        this.iItemService.delete(id);
+    @DeleteMapping(path = "/{id}")
+    public String deleteById(@PathVariable("id") String id) {
+        try {
+            if (this.iItemService.isItemExist(id)) {
+                this.iItemService.delete(id);
+                return "Item deleted successfully";
+            } else {
+                return "item could not be found";
+            }
+        }
+        catch (Exception e) {
+            return e.getMessage();
+        }
     }
     //</editor-fold>
 }
